+++ internal/repository/errors.go
@@ -5,3 +5,4 @@ import "fmt"
 var ErrDuplicateProjectName = fmt.Errorf("duplicate project name")
 var ErrDuplicateAreaName = fmt.Errorf("duplicate area name")
 var ErrDuplicateTagName = fmt.Errorf("duplicate tag name")
+var ErrSavedFilterLimitReached = fmt.Errorf("saved filter limit reached")